arch: add IDT and exceptions support
Declare and register handlers for typical x86 exceptions.
Align stack for all exception types (with and without error code) by
pushing zero when there is no error code.
Store vector number in upper 32bit of error code on the stack.

Upon exception, jump to handle_exception() routine which preserves
register context and jumps to do_exception(), which in turn panics.
The do_exception() receives a pointer to saved register context on the
stack.

Signed-off-by: Pawel Wieczorkiewicz <wipawel@amazon.de>