vmm: add initial VMM implementation
Currently only get_free_pages() and put_pages() functions are
implemented. The get_free_pages() asks PMM for a frame of specified
order and maps it to the virtual address space according to flags.
Always the highest virtual address is returned by the function.

The put_pages() unmaps just one mapping and returns the frame to PMM.
It needs to be extended to support proper unmapping, ref counting and
returning of pages only when really free.

Signed-off-by: Pawel Wieczorkiewicz <wipawel@amazon.de>