pmm: properly detect invalid paddr and mfn
Do not assume that only invalid physical addresses or frames are
only these of 0x0. To be able to correctly check if paddr or mfn
is invalid, go through all available physical memory regions and
check if paddr or mfn belongs to it. When found a matching region
the paddr/mfn is valid, otherwise invalid.

Signed-off-by: Pawel Wieczorkiewicz <wipawel@amazon.de>