Allow local overrides for build tools
Add a mechanism to the Makefile that uses an optional `Makeconf.local` file to override
variables set in the main Makefile. This mechanism along with the changes allows non-standard
overrides for all tools used in the Make targets. Also add a README section on how this
enables building and running KTF on Fedora.

Signed-off-by: Bjoern Doebel <bjoern.doebel@gmail.com>