slab allocator: added slab allocator and error codes.
Slab allocator (slab.c/slab.h) uses 8, 16, 32, 64, 128, 256, 512, 1024, 2048
slabs. As of now only 4K pages are being used. Free will be slow because there
is a search for page to which it belongs and then freed, this can be optimized
in future. It's not multi-threaded safe (so multiple tasks allocating/freeing
can corrupt links), but spinlock can be added in future.

Signed-off-by: Deepak Gupta <dkgupta@amazon.com>