Docker image: add DOCKER_MOUNT_OPTS config variable
Fedora uses some stringent SELinux configuration that fails to pass through
shared directories to a container. To fix this, we need to ask Docker to
properly adjust the directory's permissions before passing it through.
http://www.projectatomic.io/blog/2015/06/using-volumes-with-docker-can-cause-problems-with-selinux/

Add a config variable to our docker run command and extend the Fedora example
in the README.

Signed-off-by: Bjoern Doebel <bjoern.doebel@gmail.com>