compiler.h: Remove extra parentheses in call to memberof()
When switching to `-nostdinc -nostdlib` we are going to start using local
defintions of those macros that we apparently did not use before. I found a bug
in the implementation of `offsetof` that leads to a compile time error and is
fixed by removing the parentheses around the type parameter.

Signed-off-by: Bjoern Doebel <bjoern.doebel@gmail.com>