Build completely freestanding
Up to now our build pulled in the build host's type definitions from standard
library headers for C files. Avoid doing so by building ALL our files with
`-nostdinc -nostdlib`. This means we need our own implementation of a few of
the expected headers. This is mainly using GCC's builtin typedefs and mapping
them to the ppublic types. On top of that we need our own definitions of the
`NULL` and `__always_inline` macros.

Signed-off-by: Bjoern Doebel <bjoern.doebel@gmail.com>