extables: fix extables handling to handle exceptions for real
The extables handling is broken as in not being able to handle
exceptions gracefully. They'll all end up in a panic, which is probably
not intended.

Make use_extables() return a boolean that tells if the exception was
handled or not. If it was, we can simply return from the exception
handler and continue execution.

Also don't support doing both, a fixup IP and a callback. Only do one
and prefer the fixup IP which is currently the only use case.

Last, but not least, use the oppurtunity to rename the function to
extables_fixup(), as that's what it does: fixing up the exception.

Signed-off-by: Mathias Krause <minipli@grsecurity.net>