extables: introduce symbol to mark the end of defined entries
We round up the extables array to a multiple of 4kB. However, this leads to
empty extable entries, i.e. entries for NULL ptr faults without a fixup IP
nor a callback.

To avoid bogus handling of these, introduce a new symbol __stop_extables[]
that points to the end of legitimate extable entries. Leave __end_extables[]
as is to lower the amount of code churn. However, change its type to
unsigned char to avoid and flag accidental misuse.

Signed-off-by: Mathias Krause <minipli@grsecurity.net>