pmm: add merge_frames() functionality
The merge_frames() attempts to merge two consecutive frames into a higher
order frame. It only merges higher-order aligned frame with its following
frame. To do so, it checks if the frame being returned is higher-order
aligned frame and attempts to find its following frame. If the frame
being returned is not higher-order aligned, it tries to find its preceding
frame, which will be properly aligned.
When the two frames are found, the first frame's struct is reused, order
is incremented and the struct is relinked to higher order free frames list.
The second frame struct is destroyed.

The merge_frames() calls itself recursively in attempt to merge all merge-
able higher-order frames for all orders.

The put_free_frames() find corresponding frame struct for given MFN and
order. If the frame is fully returned (there is no more references) and
relinked from busy_frames to free_frames, it attempts to merge the frame.

Signed-off-by: Pawel Wieczorkiewicz <wipawel@grsecurity.net>