arch, entry: toggle kernel/user address space switch upon context switch
This uses a very simple (and just good enough for now) origin detection
mechanism. Kernel -> User or User -> Kernel tranistion is decided based
on CS selector value on the stack (CPL bits checks).

When privilege tranistion is detected a CR3 pagetable switch occurs as
well as swapgs for percpu segment (userland has it set to NULL).

User pagetables need a kernel mapping of CR3 register and percpu page
to be able to switch to it on interrupt/exception.

Signed-off-by: Pawel Wieczorkiewicz <wipawel@grsecurity.net>