cpu: don't unlink cpu after they are finished
In the wait_for_all_cpus() wait for all unfinished CPUs and/or CPUs
without empty task_queue. But do not wait for BSP as it typically
is the expected caller of this function.
Also, do not unlink CPU struct from the cpus list, there might be
more task upcoming.

Signed-off-by: Pawel Wieczorkiewicz <wipawel@grsecurity.net>