segment: restructure and document GDT selectors
The GDT entries needed to be re-organized to support sysexit. While
doing so, better document why they exist in the first place, and their
particular order.

Rename __KERN_{CS,DS}16 -> __RMODE_{CS,DS}16 to clearly indicate that
these entries only exist in the real mode GDT.

Omit entries that are unused in i386 or unused for a particular gdt.

Signed-off-by: Johannes Wikner <johannes.wikner@gmail.com>