pagetables: allow user and kernel maps in same pgd
KTF allows user mode to allocate high addresses, which normally are only
available to kernel. However, to handle this somewhat gracefully, we need
to check if already-existent page table protection bits conflicts with
those of a new pgentry.

This means that we may fixup NX, RW or USER flags if the new mapping is
more permissive than current one. This is a bit sketchy, yet useful for
certains tests. Therefore, print a big warning if this ever happens.

Signed-off-by: Johannes Wikner <johannes.wikner@gmail.com>