cmdline: add reboot_timeout option
In case communication with KTF is not possible (no PS/2 keyboard nor
UART port) add an option to automatically reboot after specified amount
of time.
The reboot_timeout by default is off. Value 0 means reboot immediately.
Values unit is seconds.

Please note a timer needs to be enabled for this functionality to work.

Signed-off-by: Pawel Wieczorkiewicz <wipawel@grsecurity.net>