arch,pt: optimize temporary mapping handling
Instead of doing full page table walk to obtain PTE of the temp.
mapping location, cache it in _tmp_mapping_entry variable.
The variable needs to be set once before construction of the page
tables begins and updated once before the switch to new page tables
(the virtual address of the PTE changes).

Also, there is no need to map in all levels of the page tables for the
_tmp_mappings, since the entry address is cached. Map only the last
level page table.

Signed-off-by: Pawel Wieczorkiewicz <wipawel@grsecurity.net>