mm,pmm: map percpu, idt and frames arrays in KERNEL_MAP area
If there is not enough early frames (or for whatever reason the frame
physical address is big enough), it might be the case that such frame
mapping does not fit into the relatively small KERNEL_BASE area.
It is better to always map in the large enough KERNEL_MAP area.

The early frames arrays (before the final page tables are created)
have to be mapped into the KERNEL_BASE area as this is the only space
covered by boot time page tables (see EARLY_VIRT_MEM settings).

Signed-off-by: Pawel Wieczorkiewicz <wipawel@grsecurity.net>