pmm: simpler and bugfix process_memory_range loop
This simplifies the adding of frames during boot. It also fixes a bug for
configurations with > 1GB memory. After the initial 4K frames, the current
(`cur`) is aligned to 2MB.  Trying to add a 1GB frame at this point will
lead to page faults further down the line as we end up with misaligned page
frame mfns.

Signed-off-by: Johannes Wikner <johannes.wikner@gmail.com>